requirements.txt
Package                Version
---------------------- -----------
aiosignal              1.2.0
anyio                  3.6.1
argon2-cffi            21.3.0
argon2-cffi-bindings   21.2.0
asttokens              2.0.5
attrs                  22.1.0
autopep8               1.6.0
Babel                  2.10.3
backcall               0.2.0
beautifulsoup4         4.11.1
bleach                 5.0.1
brotlipy               0.7.0
build                  0.8.0
certifi                2022.6.15
cffi                   1.15.0
charset-normalizer     2.0.4
click                  8.0.4
commonmark             0.9.1
conda                  4.13.0
conda-package-handling 1.8.1
cryptography           37.0.1
cycler                 0.11.0
debugpy                1.5.1
decorator              5.1.1
defusedxml             0.7.1
distlib                0.3.5
docutils               0.18.1
easydict               1.9
entrypoints            0.4
executing              0.8.3
fastjsonschema         2.16.1
filelock               3.7.1
flit_core              3.7.1
fonttools              4.34.4
frozenlist             1.3.0
grpcio                 1.43.0
idna                   3.3
imageio                2.17.0
imbalanced-learn       0.9.1
importlib-metadata     4.11.4
importlib-resources    5.9.0
iniconfig              1.1.1
ipykernel              6.9.1
ipython                8.3.0
ipython-genutils       0.2.0
jedi                   0.18.1
jeepney                0.8.0
Jinja2                 3.1.2
joblib                 1.1.0
json5                  0.9.5
jsonschema             4.9.0
jupyter-client         7.2.2
jupyter-core           4.10.0
jupyter-server         1.18.1
jupyterlab             3.4.4
jupyterlab-pygments    0.2.2
jupyterlab-server      2.15.0
keyring                23.6.0
kiwisolver             1.4.4
libauc                 1.1.9rc3
lmdb                   1.3.0
MarkupSafe             2.1.1
matplotlib             3.5.2
matplotlib-inline      0.1.2
mistune                0.8.4
mkl-fft                1.3.1
mkl-random             1.2.2
mkl-service            2.4.0
msgpack                1.0.4
nbclassic              0.4.3
nbclient               0.5.13
nbconvert              6.5.0
nbformat               5.4.0
nest-asyncio           1.5.5
networkx               2.8
notebook               6.4.12
notebook-shim          0.1.0
numpy                  1.22.3
olefile                0.46
opencv-python          4.5.5.64
packaging              21.3
pandas                 1.4.2
pandocfilters          1.5.0
parso                  0.8.3
pep517                 0.12.0
pexpect                4.8.0
pickleshare            0.7.5
Pillow                 9.1.1
pip                    21.2.4
pkginfo                1.8.3
pkgutil_resolve_name   1.3.10
platformdirs           2.5.2
pluggy                 1.0.0
prometheus-client      0.14.1
prompt-toolkit         3.0.20
protobuf               3.20.1
PTable                 0.9.2
ptyprocess             0.7.0
pure-eval              0.2.2
py                     1.11.0
pycodestyle            2.8.0
pycosat                0.6.3
pycparser              2.21
Pygments               2.12.0
pyOpenSSL              22.0.0
pyparsing              3.0.9
pyrsistent             0.18.1
PySocks                1.7.1
pytest                 7.1.2
python-dateutil        2.8.2
pytz                   2022.1
PyWavelets             1.3.0
PyYAML                 6.0
pyzmq                  22.3.0
ray                    1.13.0
readme-renderer        35.0
requests               2.28.0
requests-toolbelt      0.9.1
rfc3986                2.0.0
rich                   12.4.4
ruamel-yaml-conda      0.15.100
scikit-image           0.19.2
scikit-learn           1.1.0
scipy                  1.8.0
SecretStorage          3.3.2
Send2Trash             1.8.0
setuptools             61.2.0
six                    1.16.0
sklearn                0.0
sniffio                1.2.0
soupsieve              2.3.2.post1
stack-data             0.2.0
tabulate               0.8.10
tensorboardX           2.5.1
terminado              0.15.0
threadpoolctl          3.1.0
tifffile               2022.4.8
tinycss2               1.1.1
toml                   0.10.2
tomli                  2.0.1
torch                  1.11.0
torchaudio             0.11.0
torchvision            0.12.0
tornado                6.1
tqdm                   4.64.0
traitlets              5.1.1
twine                  4.0.1
typing_extensions      4.1.1
urllib3                1.26.9
virtualenv             20.15.1
wcwidth                0.2.5
webencodings           0.5.1
websocket-client       1.3.3
wheel                  0.37.1
XCurve                 0.0.6
zipp                   3.8.1
